misc refactors
