+++ frontend/src/app/Settings/Settings.jsx
@@ -1,27 +1,9 @@
 import React, { useState } from "react";
 import { useSelector } from "react-redux";
-import { v4 as uuidv4 } from "uuid";
 
 import DeviceSettings from "./DeviceSettings";
 import { getDevicesArray } from "../devices/deviceSelectors";
-import { DEVICE_TYPES } from "../devices/constants";
-
-const generateDeviceSettings = (devices) => {
-  return devices.reduce((acc, device) => {
-    let name = `device-${uuidv4()}`; // assign a dynamic name to each device
-    acc[name] = device;
-    return acc;
-  }, {});
-};
-
-const generateDevice = () => {
-  let deviceId = Object.keys(DEVICE_TYPES)[0]; // randomlly pick one
-  return {
-    deviceId,
-    displayName: DEVICE_TYPES[deviceId],
-    isDefault: false,
-  };
-};
+import { generateDeviceSettings } from "../devices/utils";
 
 const Settings = () => {
   const devices = useSelector(getDevicesArray);
@@ -30,14 +12,6 @@ const Settings = () => {
     generateDeviceSettings(devices)
   );
 
-  const addDevice = (deviceId) => {
-    let name = `device-${uuidv4()}`;
-    let newDeviceSettings = {
-      ...deviceSettings,
-      [name]: generateDevice(deviceId),
-    };
-    updateDeviceSettings(newDeviceSettings);
-  };
   return (
     <main className="prime-home">
       <div className="grid-container">
@@ -46,7 +20,6 @@ const Settings = () => {
       <DeviceSettings
         deviceSettings={deviceSettings}
         updateDeviceSettings={updateDeviceSettings}
-        addDevice={addDevice}
       />
     </main>
   );
